+++ R/deprec-lift.R
@@ -68,8 +68,7 @@ NULL
 #' fun <- function(x) exec("sum", !!!x)
 #' exec(sum, 3, NA, 4, na.rm = TRUE)
 lift <- function(..f, ..., .unnamed = FALSE) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift()")
+  lifecycle::deprecate_soft("1.0.0", "lift()")
 
   force(..f)
   defaults <- list(...)
@@ -88,8 +87,7 @@ lift_dl <- lift
 #' @rdname lift
 #' @export
 lift_dv <- function(..f, ..., .unnamed = FALSE) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift_dv()")
+  lifecycle::deprecate_soft("1.0.0", "lift_dv()")
 
   force(..f)
   defaults <- list(...)
@@ -123,8 +121,7 @@ lift_dv <- function(..f, ..., .unnamed = FALSE) {
 #' # now
 #' pmap_dbl(mtcars, \(...) mean(c(...)))
 lift_vl <- function(..f, ..., .type) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift_vl()")
+  lifecycle::deprecate_soft("1.0.0", "lift_vl()")
 
   force(..f)
   defaults <- list(...)
@@ -139,8 +136,7 @@ lift_vl <- function(..f, ..., .type) {
 #' @rdname lift
 #' @export
 lift_vd <- function(..f, ..., .type) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift_vd()")
+  lifecycle::deprecate_soft("1.0.0", "lift_vd()")
 
   force(..f)
   defaults <- list(...)
@@ -177,8 +173,7 @@ lift_vd <- function(..f, ..., .type) {
 #' mtcars |> pmap_lgl(\(...) any(c(...) > 200))
 #'
 lift_ld <- function(..f, ...) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift_ld()")
+  lifecycle::deprecate_soft("1.0.0", "lift_ld()")
 
   force(..f)
   defaults <- list(...)
@@ -190,8 +185,7 @@ lift_ld <- function(..f, ...) {
 #' @rdname lift
 #' @export
 lift_lv <- function(..f, ...) {
-  # in 1.0.0
-  lifecycle::signal_stage("deprecated", "lift_lv()")
+  lifecycle::deprecate_soft("1.0.0", "lift_lv()")
 
   force(..f)
   defaults <- list(...)